backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
mysqlclient==2.2.7
pip==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.11.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
wheel==0.45.1
